Updated brush::impl_trait macro. Now it allow to pass the ident of trait with implementation. Added brush::trait_definition to hide reflaction functionality. Removed underscore from all public methods inside of internal trait. Removed Erc*Internal traits. All methods from them moved to according public trait. Simplified public functions to not return result with error. Now all contracts will assert if some inner check can't be pass. Moved unit tests to separate folders.
