+++ contracts/access/access-control/lib.rs
@@ -2,3 +2,4 @@
 
 pub mod impls;
 pub mod traits;
+mod tests;
