+++ contracts/access/ownable/impls.rs
@@ -1,4 +1,4 @@
-use utils::{
+use brush::{
     traits::{InkStorage, AccountId},
     define_getters,
 };
@@ -8,38 +8,33 @@ pub trait OwnableStorage: InkStorage {
     define_getters!(_owner, _owner_mut, AccountId);
 }
 
-pub trait OwnableModifiers: OwnableStorage {
-    fn only_owner(&self) -> Result<(), OwnableError> {
-        if self._owner() != &Self::env().caller() {
-            return Err(OwnableError::CallerIsNotOwner)
-        }
-        Ok(())
-    }
-}
-
 const ZERO_ADDRESS: [u8; 32] = [0; 32];
 
-pub trait Ownable: OwnableModifiers {
+pub trait Ownable: OwnableStorage {
+    fn only_owner(&self) {
+        assert_eq!(self._owner(), &Self::env().caller(), "{}", OwnableError::CallerIsNotOwner.as_ref());
+    }
+
     fn _init_with_owner(&mut self, owner: AccountId) {
         *self._owner_mut() = owner;
         // TODO: Emit event
     }
 
-    fn _renounce_ownership(&mut self) -> Result<(), OwnableError> {
-        self.only_owner()?;
+    fn owner(&self) -> AccountId {
+        self._owner().clone()
+    }
+
+    fn renounce_ownership(&mut self) {
+        self.only_owner();
 
         // TODO: Emit event
         *self._owner_mut() = ZERO_ADDRESS.into();
-        Ok(())
     }
 
-    fn _transfer_ownership(&mut self, new_owner: AccountId) -> Result<(), OwnableError> {
-        self.only_owner()?;
-        if new_owner == ZERO_ADDRESS.into() {
-            return Err(OwnableError::NewOwnerIsZero)
-        }
+    fn transfer_ownership(&mut self, new_owner: AccountId) {
+        self.only_owner();
+        assert_ne!(new_owner, ZERO_ADDRESS.into(), "{}", OwnableError::NewOwnerIsZero.as_ref());
         // TODO: Emit event
         *self._owner_mut() = new_owner;
-        Ok(())
     }
 }
