+++ contracts/brush/proc_macros/lib.rs
@@ -1,4 +1,3 @@
-#![cfg_attr(not(feature = "std"), no_std)]
 extern crate proc_macro;
 
 use quote::{
@@ -10,6 +9,7 @@ use syn::{
     ItemTrait,
     Item,
     parse_macro_input,
+    parse::Parser,
 };
 use proc_macro::TokenStream;
 use proc_macro2::{
@@ -27,8 +27,9 @@ const TEMP_FILE: &str = "brush_temp";
 type Data = HashMap<String, Vec<String>>;
 
 #[proc_macro_attribute]
-pub fn contract(_: TokenStream, ink_module: TokenStream) -> TokenStream {
+pub fn contract(_attrs: TokenStream, ink_module: TokenStream) -> TokenStream {
     let input: TokenStream2 = ink_module.into();
+    let attrs: TokenStream2 = _attrs.into();
     let mut module = syn::parse2::<syn::ItemMod>(input.clone()).unwrap();
     let (braces, items) = match module.content {
         Some((brace, items)) => (brace, items),
@@ -39,51 +40,113 @@ pub fn contract(_: TokenStream, ink_module: TokenStream) -> TokenStream {
         }
     };
 
-    let items = items
+    let mut new_items: Vec<syn::Item> = vec![];
+    let mut items: Vec<syn::Item> = items
         .into_iter()
-        .map(|item| {
+        .filter_map(|item| {
             if let Item::Macro(macro_item) = &item {
-                if macro_item.mac.path.is_ident("make_trait") {
-                    syn::parse2::<syn::Item>(
-                        make_trait(macro_item.mac.tokens.clone().into())
-                            .into()
-                    ).unwrap()
+                if macro_item.mac.path.is_ident("impl_trait") {
+                    // let stream: TokenStream2 = impl_trait(macro_item.mac.tokens.clone().into()).into();
+                    // stream.
+                    let mut generated_items = syn::Block::parse_within.
+                        parse(impl_trait(macro_item.mac.tokens.clone().into()))
+                        .unwrap()
+                        .iter_mut()
+                        .filter_map(|stmt|
+                            if let syn::Stmt::Item(item) = stmt {
+                                Some(item.clone())
+                            } else {
+                                None
+                            }
+                        ).collect();
+                    new_items.append(&mut generated_items);
+                    None
                 } else {
-                    item
+                    Some(item)
                 }
             } else {
-                item
+                Some(item)
             }
         }).collect();
+
+    items.append(&mut new_items);
     module.content = Some((braces, items));
 
     let result = quote! {
+        #attrs
         #[ink_lang::contract]
         #module
     };
     result.into()
 }
 
+#[proc_macro_attribute]
+pub fn trait_definition(_attrs: TokenStream, _input: TokenStream) -> TokenStream {
+    let input: TokenStream2 = _input.clone().into();
+    let attrs: TokenStream2 = _attrs.into();
+    let trait_item = parse_macro_input!(_input as ItemTrait);
+
+    let mut hash_map = load_hash_map();
+    put_trait(&mut hash_map, trait_item);
+    save_hash_map(hash_map);
+
+    let code = quote! {
+        #attrs
+        #[ink_lang::trait_definition]
+        #input
+    };
+    code.into()
+}
+
+struct ImplTrait {
+    contract: syn::Ident,
+    traits: Vec<(syn::Ident, Option<syn::Ident>)>,
+}
+
+fn parse_impl_trait(stream: TokenStream2) -> ImplTrait {
+    let mut iter = stream.into_iter();
+    let mut impl_trait;
+    if let TokenTree::Ident(ident) = iter.next().unwrap() {
+        impl_trait = ImplTrait{
+            contract: ident,
+            traits: vec![]
+        };
+    } else {
+        panic!("First token is not struct ident");
+    }
+    for item in iter {
+        if let TokenTree::Punct(_) = item {
+            continue
+        } else if let TokenTree::Ident(trait_ident) = item {
+            impl_trait.traits.push((trait_ident, None));
+        } else if let TokenTree::Group(trait_group) = item {
+            if let Some(pair) = impl_trait.traits.last_mut() {
+                let internal_trait = &mut pair.1;
+                if internal_trait.is_none() {
+                    *internal_trait = Some(syn::parse2::<syn::Ident>(trait_group.stream()).unwrap());
+                } else {
+                    panic!("External trait already contains internal");
+                }
+            } else {
+                panic!("Internal trait before external");
+            }
+        }
+    }
+    impl_trait
+}
+
 #[proc_macro]
-pub fn make_trait(_item: TokenStream) -> TokenStream {
-    // println!("make_trait {}", format!("{:?}", _item));
-    let mut tokens: Vec<syn::Ident> = TokenStream2::from(_item)
-        .into_iter().filter_map(|item|
-        if let TokenTree::Ident(ident) = item {
-            Some(ident)
-        } else {
-            None
-        }).collect();
+pub fn impl_trait(_item: TokenStream) -> TokenStream {
+    let impl_trait = parse_impl_trait(TokenStream2::from(_item));
 
-    let contract_ident = tokens.drain(0..1).last().unwrap();
+    let contract_ident = impl_trait.contract;
 
-    let contract_traits = tokens;
     let traits = load_hash_map();
     let mut impls: Vec<TokenStream2> = vec![];
 
-    for trait_ident in contract_traits.iter() {
-        // println!("{:?}", trait_ident);
-        let trait_methods: Vec<_> = traits.get(&trait_ident.to_string()).unwrap().into_iter()
+    // println!("{:?}", impl_trait.clone());
+    for (external_trait, internal) in impl_trait.traits.iter() {
+        let trait_methods: Vec<_> = traits.get(&external_trait.to_string()).unwrap().into_iter()
             .map(|method| TokenStream2::from_str(method).unwrap())
             .map(|stream| syn::parse2::<syn::TraitItemMethod>(stream).unwrap())
             .collect();
@@ -113,21 +176,26 @@ pub fn make_trait(_item: TokenStream) -> TokenStream {
                 }
                 _ => unreachable!("encountered invalid receiver argument for brush message"),
             };
-            // transform IErc20 -> Erc20
-            let mut chars = trait_ident.to_string().clone();
-            chars.remove(0);
-            let impl_trait_ident = format_ident!("{}", chars.as_str());
+            let internal_trait;
+            if let Some(t) = internal {
+                internal_trait = t.clone();
+            } else {
+                // transform IErc20 -> Erc20
+                let mut chars = external_trait.to_string().clone();
+                chars.remove(0);
+                internal_trait = format_ident!("{}", chars.as_str());
+            }
 
             let output = item.sig.output;
             quote! {
                 #( #attrs )*
                 fn #ident( #receiver #(, #inputs )* ) #output {
-                    <#contract_ident as #impl_trait_ident>::#ident(self #(,  #inputs_params )*)
+                    <#contract_ident as #internal_trait>::#ident(self #(,  #inputs_params )*)
                 }
             }
         });
         let gen = quote! {
-            impl #trait_ident for #contract_ident {
+            impl #external_trait for #contract_ident {
                 #(#implementations)*
             }
         };
@@ -139,19 +207,6 @@ pub fn make_trait(_item: TokenStream) -> TokenStream {
         #(#impls_iter)*
     };
     final_code.into()
-}
-
-#[proc_macro_attribute]
-pub fn reflect_trait(_: TokenStream, input: TokenStream) -> TokenStream {
-    // println!("reflect_trait {}", format!("{:?}", input));
-    let trait_input = input.clone();
-    let trait_item = parse_macro_input!(trait_input as ItemTrait);
-
-    let mut hash_map = load_hash_map();
-    put_trait(&mut hash_map, trait_item);
-    save_hash_map(hash_map);
-
-    input
 }
 
 fn load_hash_map() -> Data {
@@ -173,7 +228,6 @@ fn save_hash_map(hashmap: Data) {
     let mut dir = env::temp_dir();
     dir = dir.join(TEMP_FILE);
 
-    println!("Save {}", format!("{:?}", dir));
     let file = OpenOptions::new().write(true).truncate(true).open(&dir).unwrap();
     serde_json::to_writer(file, &hashmap).unwrap();
 }
