+++ contracts/brush/traits.rs
