+++ contracts/token/erc20/impls.rs
@@ -1,17 +1,15 @@
-use crate::traits::{
-    Erc20Error,
-};
+use crate::traits::Erc20Error;
 use ink_storage::{
     collections::{
         HashMap as StorageHashMap,
     },
     Lazy,
 };
-use utils::{
+use brush::{
     traits::{InkStorage, AccountId, Balance},
     define_getters,
 };
-use ink_prelude::{string::String};
+use ink_prelude::{string::{String}};
 
 const ZERO_ADDRESS: [u8; 32] = [0; 32];
 
@@ -24,60 +22,13 @@ pub trait Erc20Storage: InkStorage {
     define_getters!(_decimals, _decimals_mut, Lazy<u8>);
 }
 
-pub trait Erc20Internal: Erc20Storage {
-    fn _emit_transfer_event(&self, _from: Option<AccountId>, _to: Option<AccountId>, _amount: Balance);
+pub trait Erc20: Erc20Storage {
+    /// Emit transfer event. It must be implemented in inherited struct
+    fn emit_transfer_event(&self, _from: Option<AccountId>, _to: Option<AccountId>, _amount: Balance);
 
-    fn _emit_approval_event(&self, _owner: AccountId, _spender: AccountId, _amount: Balance);
+    /// Emit approval event. It must be implemented in inherited struct
+    fn emit_approval_event(&self, _owner: AccountId, _spender: AccountId, _amount: Balance);
 
-    fn _balance_of(&self, owner: AccountId) -> Balance {
-        self._balances().get(&owner).copied().unwrap_or(0)
-    }
-
-    fn _before_token_transfer(&mut self, _from: AccountId, _to: AccountId, _amount: Balance) -> Result<(), Erc20Error> {
-        Ok(())
-    }
-
-    fn _transfer_from_to(
-        &mut self,
-        from: AccountId,
-        to: AccountId,
-        amount: Balance,
-    ) -> Result<(), Erc20Error> {
-        if from == ZERO_ADDRESS.into() {
-            return Err(Erc20Error::ZeroSenderAddress);
-        }
-        if to == ZERO_ADDRESS.into() {
-            return Err(Erc20Error::ZeroRecipientAddress);
-        }
-
-        self._before_token_transfer(from, to, amount)?;
-
-        let from_balance = self._balance_of(from);
-        if from_balance < amount {
-            return Err(Erc20Error::InsufficientBalance);
-        }
-        self._balances_mut().insert(from, from_balance - amount);
-        let to_balance = self._balance_of(to);
-        self._balances_mut().insert(to, to_balance + amount);
-        self._emit_transfer_event(Some(from), Some(to), amount);
-        Ok(())
-    }
-
-    fn _approve_from_to(&mut self, owner: AccountId, spender: AccountId, amount: Balance) -> Result<(), Erc20Error> {
-        if owner == ZERO_ADDRESS.into() {
-            return Err(Erc20Error::ZeroSenderAddress);
-        }
-        if spender == ZERO_ADDRESS.into() {
-            return Err(Erc20Error::ZeroRecipientAddress);
-        }
-
-        self._allowances_mut().insert((owner, spender), amount);
-        self._emit_approval_event(owner, spender, amount);
-        Ok(())
-    }
-}
-
-pub trait Erc20: Erc20Internal {
     /// Returns the token name.
     fn token_name(&self) -> Option<String> {
         Lazy::get(self._name()).clone()
@@ -102,7 +53,7 @@ pub trait Erc20: Erc20Internal {
     ///
     /// Returns `0` if the account is non-existent.
     fn balance_of(&self, owner: AccountId) -> Balance {
-        self._balance_of(owner)
+        self._balances().get(&owner).copied().unwrap_or(0)
     }
 
     /// Transfers `value` amount of tokens from the caller's account to account `to`.
@@ -111,13 +62,13 @@ pub trait Erc20: Erc20Internal {
     ///
     /// # Errors
     ///
-    /// Returns `InsufficientBalance` error if there are not enough tokens on
+    /// Panics `InsufficientBalance` error if there are not enough tokens on
     /// the caller's account Balance.
     ///
-    /// Returns `ZeroSenderAddress` error if sender's address is zero.
+    /// Panics `ZeroSenderAddress` error if sender's address is zero.
     ///
-    /// Returns `ZeroRecipientAddress` error if recipient's address is zero.
-    fn transfer(&mut self, to: AccountId, value: Balance) -> Result<(), Erc20Error> {
+    /// Panics `ZeroRecipientAddress` error if recipient's address is zero.
+    fn transfer(&mut self, to: AccountId, value: Balance) {
         let from = Self::env().caller();
         self._transfer_from_to(from, to, value)
     }
@@ -138,24 +89,21 @@ pub trait Erc20: Erc20Internal {
     ///
     /// # Errors
     ///
-    /// Returns `InsufficientAllowance` error if there are not enough tokens allowed
+    /// Panics `InsufficientAllowance` error if there are not enough tokens allowed
     /// for the caller to withdraw from `from`.
     ///
-    /// Returns `InsufficientBalance` error if there are not enough tokens on
+    /// Panics `InsufficientBalance` error if there are not enough tokens on
     /// the the account Balance of `from`.
     ///
-    /// Returns `ZeroSenderAddress` error if sender's address is zero.
+    /// Panics `ZeroSenderAddress` error if sender's address is zero.
     ///
-    /// Returns `ZeroRecipientAddress` error if recipient's address is zero.
-    fn transfer_from(&mut self, from: AccountId, to: AccountId, value: Balance) -> Result<(), Erc20Error> {
+    /// Panics `ZeroRecipientAddress` error if recipient's address is zero.
+    fn transfer_from(&mut self, from: AccountId, to: AccountId, value: Balance) {
         let caller = Self::env().caller();
         let allowance = self.allowance(from, caller);
-        if allowance < value {
-            return Err(Erc20Error::InsufficientAllowance);
-        }
-        self._transfer_from_to(from, to, value)?;
-        self._approve_from_to(from, caller, allowance - value)?;
-        Ok(())
+        assert!(allowance >= value, "{}", Erc20Error::InsufficientAllowance.as_ref());
+        self._transfer_from_to(from, to, value);
+        self._approve_from_to(from, caller, allowance - value);
     }
 
     /// Allows `spender` to withdraw from the caller's account multiple times, up to
@@ -167,10 +115,10 @@ pub trait Erc20: Erc20Internal {
     ///
     /// # Errors
     ///
-    /// Returns `ZeroSenderAddress` error if sender's address is zero.
+    /// Panics `ZeroSenderAddress` error if sender's address is zero.
     ///
-    /// Returns `ZeroRecipientAddress` error if recipient's address is zero.
-    fn approve(&mut self, spender: AccountId, value: Balance) -> Result<(), Erc20Error> {
+    /// Panics `ZeroRecipientAddress` error if recipient's address is zero.
+    fn approve(&mut self, spender: AccountId, value: Balance) {
         let owner = Self::env().caller();
         self._approve_from_to(owner, spender, value)
     }
@@ -186,10 +134,10 @@ pub trait Erc20: Erc20Internal {
     ///
     /// # Errors
     ///
-    /// Returns `ZeroSenderAddress` error if sender's address is zero.
+    /// Panics `ZeroSenderAddress` error if sender's address is zero.
     ///
-    /// Returns `ZeroRecipientAddress` error if recipient's address is zero.
-    fn increase_allowance(&mut self, spender: AccountId, delta_value: Balance) -> Result<(), Erc20Error> {
+    /// Panics `ZeroRecipientAddress` error if recipient's address is zero.
+    fn increase_allowance(&mut self, spender: AccountId, delta_value: Balance) {
         let owner = Self::env().caller();
         self._approve_from_to(owner, spender, self.allowance(owner, spender) + delta_value)
     }
@@ -200,18 +148,16 @@ pub trait Erc20: Erc20Internal {
     ///
     /// # Errors
     ///
-    /// Returns `InsufficientAllowance` error if there are not enough tokens allowed
+    /// Panics `InsufficientAllowance` error if there are not enough tokens allowed
     /// by owner for `spender`.
     ///
-    /// Returns `ZeroSenderAddress` error if sender's address is zero.
+    /// Panics `ZeroSenderAddress` error if sender's address is zero.
     ///
-    /// Returns `ZeroRecipientAddress` error if recipient's address is zero.
-    fn decrease_allowance(&mut self, spender: AccountId, delta_value: Balance) -> Result<(), Erc20Error> {
+    /// Panics `ZeroRecipientAddress` error if recipient's address is zero.
+    fn decrease_allowance(&mut self, spender: AccountId, delta_value: Balance) {
         let owner = Self::env().caller();
         let allowance = self.allowance(owner, spender);
-        if allowance < delta_value {
-            return Err(Erc20Error::InsufficientAllowance);
-        }
+        assert!(allowance >= delta_value, "{}", Erc20Error::InsufficientAllowance.as_ref());
 
         self._approve_from_to(owner, spender, allowance - delta_value)
     }
@@ -222,18 +168,15 @@ pub trait Erc20: Erc20Internal {
     ///
     /// # Errors
     ///
-    /// Returns `ZeroRecipientAddress` error if recipient's address is zero.
-    fn mint(&mut self, account: AccountId, amount: Balance) -> Result<(), Erc20Error> {
-        if account == ZERO_ADDRESS.into() {
-            return Err(Erc20Error::ZeroRecipientAddress);
-        }
+    /// Panics `ZeroRecipientAddress` error if recipient's address is zero.
+    fn mint(&mut self, account: AccountId, amount: Balance) {
+        assert!(account != ZERO_ADDRESS.into(), "{}", Erc20Error::ZeroRecipientAddress.as_ref());
 
-        let mut new_balance = self._balance_of(account);
+        let mut new_balance = self.balance_of(account);
         new_balance += amount;
         self._balances_mut().insert(account, new_balance);
         *self._supply_mut() = Lazy::new(self.total_supply() + amount);
-        self._emit_transfer_event(None, Some(account), amount);
-        Ok(())
+        self.emit_transfer_event(None, Some(account), amount);
     }
 
     /// Destroys `amount` tokens from `account`, reducing the total supply.
@@ -242,445 +185,45 @@ pub trait Erc20: Erc20Internal {
     ///
     /// # Errors
     ///
-    /// Returns `ZeroSenderAddress` error if recipient's address is zero.
+    /// Panics `ZeroSenderAddress` error if recipient's address is zero.
     ///
-    /// Returns `InsufficientBalance` error if there are not enough tokens on
+    /// Panics `InsufficientBalance` error if there are not enough tokens on
     /// the the account Balance of `account`.
-    fn burn(&mut self, account: AccountId, amount: Balance) -> Result<(), Erc20Error> {
-        if account == ZERO_ADDRESS.into() {
-            return Err(Erc20Error::ZeroSenderAddress);
-        }
+    fn burn(&mut self, account: AccountId, amount: Balance) {
+        assert!(account != ZERO_ADDRESS.into(), "{}", Erc20Error::ZeroSenderAddress.as_ref());
 
-        let mut from_balance = self._balance_of(account);
-        if from_balance < amount {
-            return Err(Erc20Error::InsufficientBalance);
-        }
+        let mut from_balance = self.balance_of(account);
+        assert!(from_balance >= amount, "{}", Erc20Error::InsufficientBalance.as_ref());
 
         from_balance -= amount;
         self._balances_mut().insert(account, from_balance);
         *self._supply_mut() = Lazy::new(self.total_supply() - amount);
-        self._emit_transfer_event(Some(account), None, amount);
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-#[ink_lang::contract]
-mod tests {
-    /// Imports all the definitions from the outer scope so we can use them here.
-    use super::*;
-    use ink_lang as ink;
-    use utils::{
-        traits::{InkStorage},
-        iml_getters,
-    };
-    use ink::{Env, EmitEvent};
-    use crate::traits::{ IErc20 };
-    use ink_env::{
-        hash::{
-            Blake2x256,
-            CryptoHash,
-            HashOutput,
-        },
-        Clear,
-    };
-
-    /// Event emitted when a token transfer occurs.
-    #[ink(event)]
-    pub struct Transfer {
-        #[ink(topic)]
-        from: Option<AccountId>,
-        #[ink(topic)]
-        to: Option<AccountId>,
-        value: Balance,
+        self.emit_transfer_event(Some(account), None, amount);
     }
 
-    /// Event emitted when an approval occurs that `spender` is allowed to withdraw
-    /// up to the amount of `value` tokens from `owner`.
-    #[ink(event)]
-    pub struct Approval {
-        #[ink(topic)]
-        owner: AccountId,
-        #[ink(topic)]
-        spender: AccountId,
-        value: Balance,
+    // Internal functions
+
+    fn _before_token_transfer(&mut self, _from: AccountId, _to: AccountId, _amount: Balance) {}
+
+    fn _transfer_from_to(&mut self, from: AccountId, to: AccountId, amount: Balance) {
+        assert!(from != ZERO_ADDRESS.into(), "{}", Erc20Error::ZeroSenderAddress.as_ref());
+        assert!(to != ZERO_ADDRESS.into(), "{}", Erc20Error::ZeroRecipientAddress.as_ref());
+
+        self._before_token_transfer(from, to, amount);
+
+        let from_balance = self.balance_of(from);
+        assert!(from_balance >= amount, "{}", Erc20Error::InsufficientBalance.as_ref());
+        self._balances_mut().insert(from, from_balance - amount);
+        let to_balance = self.balance_of(to);
+        self._balances_mut().insert(to, to_balance + amount);
+        self.emit_transfer_event(Some(from), Some(to), amount);
     }
 
-    /// A simple ERC-20 contract.
-    #[ink(storage)]
-    #[derive(Default)]
-    pub struct Erc20Struct {
-        total_supply: Lazy<Balance>,
-        balances: StorageHashMap<AccountId, Balance>,
-        allowances: StorageHashMap<(AccountId, AccountId), Balance>,
-        name: Lazy<Option<String>>,
-        symbol: Lazy<Option<String>>,
-        decimal: Lazy<u8>,
-    }
-    type Event = <Erc20Struct as ::ink_lang::BaseEvent>::Type;
+    fn _approve_from_to(&mut self, owner: AccountId, spender: AccountId, amount: Balance) {
+        assert!(owner != ZERO_ADDRESS.into(), "{}", Erc20Error::ZeroSenderAddress.as_ref());
+        assert!(spender != ZERO_ADDRESS.into(), "{}", Erc20Error::ZeroRecipientAddress.as_ref());
 
-    impl InkStorage for Erc20Struct {}
-    impl Erc20Storage for Erc20Struct {
-        iml_getters!(total_supply, _supply, _supply_mut, Lazy<Balance>);
-        iml_getters!(balances, _balances, _balances_mut, StorageHashMap<AccountId, Balance>);
-        iml_getters!(allowances, _allowances, _allowances_mut, StorageHashMap<(AccountId, AccountId), Balance>);
-        iml_getters!(name, _name, _name_mut, Lazy<Option<String>>);
-        iml_getters!(symbol, _symbol, _symbol_mut, Lazy<Option<String>>);
-        iml_getters!(decimal, _decimals, _decimals_mut, Lazy<u8>);
-    }
-    impl Erc20Internal for Erc20Struct {
-        fn _emit_transfer_event(&self, _from: Option<AccountId>, _to: Option<AccountId>, _amount: Balance) {
-            self.env().emit_event(Transfer {
-                from: _from,
-                to: _to,
-                value: _amount,
-            });
-        }
-
-        fn _emit_approval_event(&self, _owner: AccountId, _spender: AccountId, _amount: Balance) {
-            self.env().emit_event(Approval {
-                owner: _owner,
-                spender: _spender,
-                value: _amount,
-            });
-        }
-    }
-    impl Erc20 for Erc20Struct {}
-    
-    impl Erc20Struct {
-        #[ink(constructor)]
-        pub fn new(_total_supply: Balance) -> Self {
-            let mut instance = Self::_empty();
-            instance.mint(instance.env().caller(), _total_supply).expect("Can't mint tokens");
-            instance
-        }
-    }
-
-    fn assert_transfer_event(
-        event: &ink_env::test::EmittedEvent,
-        expected_from: Option<AccountId>,
-        expected_to: Option<AccountId>,
-        expected_value: Balance,
-    ) {
-        let decoded_event = <Event as scale::Decode>::decode(&mut &event.data[..])
-            .expect("encountered invalid contract event data buffer");
-        if let Event::Transfer(Transfer { from, to, value }) = decoded_event {
-            assert_eq!(from, expected_from, "encountered invalid Transfer.from");
-            assert_eq!(to, expected_to, "encountered invalid Transfer.to");
-            assert_eq!(value, expected_value, "encountered invalid Trasfer.value");
-        } else {
-            panic!("encountered unexpected event kind: expected a Transfer event")
-        }
-        fn encoded_into_hash<T>(entity: &T) -> Hash
-            where
-                T: scale::Encode,
-        {
-            let mut result = Hash::clear();
-            let len_result = result.as_ref().len();
-            let encoded = entity.encode();
-            let len_encoded = encoded.len();
-            if len_encoded <= len_result {
-                result.as_mut()[..len_encoded].copy_from_slice(&encoded);
-                return result
-            }
-            let mut hash_output =
-                <<Blake2x256 as HashOutput>::Type as Default>::default();
-            <Blake2x256 as CryptoHash>::hash(&encoded, &mut hash_output);
-            let copy_len = core::cmp::min(hash_output.len(), len_result);
-            result.as_mut()[0..copy_len].copy_from_slice(&hash_output[0..copy_len]);
-            result
-        }
-        let expected_topics = vec![
-            encoded_into_hash(&PrefixedValue {
-                value: b"Erc20Struct::Transfer",
-                prefix: b"",
-            }),
-            encoded_into_hash(&PrefixedValue {
-                prefix: b"Erc20Struct::Transfer::from",
-                value: &expected_from,
-            }),
-            encoded_into_hash(&PrefixedValue {
-                prefix: b"Erc20Struct::Transfer::to",
-                value: &expected_to,
-            }),
-            encoded_into_hash(&PrefixedValue {
-                prefix: b"Erc20Struct::Transfer::value",
-                value: &expected_value,
-            }),
-        ];
-        for (n, (actual_topic, expected_topic)) in
-        event.topics.iter().zip(expected_topics).enumerate()
-        {
-            let topic = actual_topic
-                .decode::<Hash>()
-                .expect("encountered invalid topic encoding");
-            assert_eq!(topic, expected_topic, "encountered invalid topic at {}", n);
-        }
-    }
-
-    /// The default constructor does its job.
-    #[ink::test]
-    fn new_works() {
-        // Constructor works.
-        let _erc20 = Erc20Struct::new(100);
-
-        // Transfer event triggered during initial construction.
-        let emitted_events = ink_env::test::recorded_events().collect::<Vec<_>>();
-        assert_eq!(1, emitted_events.len());
-
-        assert_transfer_event(
-            &emitted_events[0],
-            None,
-            Some(AccountId::from([0x01; 32])),
-            100,
-        );
-    }
-
-    /// The total supply was applied.
-    #[ink::test]
-    fn total_supply_works() {
-        // Constructor works.
-        let erc20 = Erc20Struct::new(100);
-        // Transfer event triggered during initial construction.
-        let emitted_events = ink_env::test::recorded_events().collect::<Vec<_>>();
-        assert_transfer_event(
-            &emitted_events[0],
-            None,
-            Some(AccountId::from([0x01; 32])),
-            100,
-        );
-        // Get the token total supply.
-        assert_eq!(erc20.total_supply(), 100);
-    }
-
-    /// Get the actual balance of an account.
-    #[ink::test]
-    fn balance_of_works() {
-        // Constructor works
-        let erc20 = Erc20Struct::new(100);
-        // Transfer event triggered during initial construction
-        let emitted_events = ink_env::test::recorded_events().collect::<Vec<_>>();
-        assert_transfer_event(
-            &emitted_events[0],
-            None,
-            Some(AccountId::from([0x01; 32])),
-            100,
-        );
-        let accounts =
-            ink_env::test::default_accounts::<ink_env::DefaultEnvironment>()
-                .expect("Cannot get accounts");
-        // Alice owns all the tokens on deployment
-        assert_eq!(erc20.balance_of(accounts.alice), 100);
-        // Bob does not owns tokens
-        assert_eq!(erc20.balance_of(accounts.bob), 0);
-    }
-
-    #[ink::test]
-    fn transfer_works() {
-        // Constructor works.
-        let mut erc20 = Erc20Struct::new(100);
-        // Transfer event triggered during initial construction.
-        let accounts =
-            ink_env::test::default_accounts::<ink_env::DefaultEnvironment>()
-                .expect("Cannot get accounts");
-
-        assert_eq!(erc20.balance_of(accounts.bob), 0);
-        // Alice transfers 10 tokens to Bob.
-        assert_eq!(erc20.transfer(accounts.bob, 10), Ok(()));
-        // Bob owns 10 tokens.
-        assert_eq!(erc20.balance_of(accounts.bob), 10);
-
-        let emitted_events = ink_env::test::recorded_events().collect::<Vec<_>>();
-        assert_eq!(emitted_events.len(), 2);
-        // Check first transfer event related to ERC-20 instantiation.
-        assert_transfer_event(
-            &emitted_events[0],
-            None,
-            Some(AccountId::from([0x01; 32])),
-            100,
-        );
-        // Check the second transfer event relating to the actual trasfer.
-        assert_transfer_event(
-            &emitted_events[1],
-            Some(AccountId::from([0x01; 32])),
-            Some(AccountId::from([0x02; 32])),
-            10,
-        );
-    }
-
-    #[ink::test]
-    fn invalid_transfer_should_fail() {
-        // Constructor works.
-        let mut erc20 = Erc20Struct::new(100);
-        let accounts =
-            ink_env::test::default_accounts::<ink_env::DefaultEnvironment>()
-                .expect("Cannot get accounts");
-
-        assert_eq!(erc20.balance_of(accounts.bob), 0);
-        // Get contract address.
-        let callee = ink_env::account_id::<ink_env::DefaultEnvironment>()
-            .unwrap_or([0x0; 32].into());
-        // Create call
-        let mut data =
-            ink_env::test::CallData::new(ink_env::call::Selector::new([0x00; 4])); // balance_of
-        data.push_arg(&accounts.bob);
-        // Push the new execution context to set Bob as caller
-        ink_env::test::push_execution_context::<ink_env::DefaultEnvironment>(
-            accounts.bob,
-            callee,
-            1000000,
-            1000000,
-            data,
-        );
-
-        // Bob fails to transfers 10 tokens to Eve.
-        assert_eq!(
-            erc20.transfer(accounts.eve, 10),
-            Err(Erc20Error::InsufficientBalance)
-        );
-        // Alice owns all the tokens.
-        assert_eq!(erc20.balance_of(accounts.alice), 100);
-        assert_eq!(erc20.balance_of(accounts.bob), 0);
-        assert_eq!(erc20.balance_of(accounts.eve), 0);
-
-        // Transfer event triggered during initial construction.
-        let emitted_events = ink_env::test::recorded_events().collect::<Vec<_>>();
-        assert_eq!(emitted_events.len(), 1);
-        assert_transfer_event(
-            &emitted_events[0],
-            None,
-            Some(AccountId::from([0x01; 32])),
-            100,
-        );
-    }
-
-    #[ink::test]
-    fn transfer_from_works() {
-        // Constructor works.
-        let mut erc20 = Erc20Struct::new(100);
-        // Transfer event triggered during initial construction.
-        let accounts =
-            ink_env::test::default_accounts::<ink_env::DefaultEnvironment>()
-                .expect("Cannot get accounts");
-
-        // Bob fails to transfer tokens owned by Alice.
-        assert_eq!(
-            erc20.transfer_from(accounts.alice, accounts.eve, 10),
-            Err(Erc20Error::InsufficientAllowance)
-        );
-        // Alice approves Bob for token transfers on her behalf.
-        assert_eq!(erc20.approve(accounts.bob, 10), Ok(()));
-
-        // The approve event takes place.
-        assert_eq!(ink_env::test::recorded_events().count(), 2);
-
-        // Get contract address.
-        let callee = ink_env::account_id::<ink_env::DefaultEnvironment>()
-            .unwrap_or([0x0; 32].into());
-        // Create call.
-        let mut data =
-            ink_env::test::CallData::new(ink_env::call::Selector::new([0x00; 4])); // balance_of
-        data.push_arg(&accounts.bob);
-        // Push the new execution context to set Bob as caller.
-        ink_env::test::push_execution_context::<ink_env::DefaultEnvironment>(
-            accounts.bob,
-            callee,
-            1000000,
-            1000000,
-            data,
-        );
-
-        // Bob transfers tokens from Alice to Eve.
-        assert_eq!(
-            erc20.transfer_from(accounts.alice, accounts.eve, 10),
-            Ok(())
-        );
-        // Eve owns tokens.
-        assert_eq!(erc20.balance_of(accounts.eve), 10);
-
-        // Check all transfer events that happened during the previous calls:
-        let emitted_events = ink_env::test::recorded_events().collect::<Vec<_>>();
-        assert_eq!(emitted_events.len(), 4);
-        assert_transfer_event(
-            &emitted_events[0],
-            None,
-            Some(AccountId::from([0x01; 32])),
-            100,
-        );
-        // The second event `emitted_events[1]` is an Approve event that we skip checking.
-        assert_transfer_event(
-            &emitted_events[2],
-            Some(AccountId::from([0x01; 32])),
-            Some(AccountId::from([0x05; 32])),
-            10,
-        );
-    }
-
-    #[ink::test]
-    fn allowance_must_not_change_on_failed_transfer() {
-        let mut erc20 = Erc20Struct::new(100);
-        let accounts =
-            ink_env::test::default_accounts::<ink_env::DefaultEnvironment>()
-                .expect("Cannot get accounts");
-
-        // Alice approves Bob for token transfers on her behalf.
-        let alice_balance = erc20.balance_of(accounts.alice);
-        let initial_allowance = alice_balance + 2;
-        assert_eq!(erc20.approve(accounts.bob, initial_allowance), Ok(()));
-
-        // Get contract address.
-        let callee = ink_env::account_id::<ink_env::DefaultEnvironment>()
-            .unwrap_or([0x0; 32].into());
-        // Create call.
-        let mut data =
-            ink_env::test::CallData::new(ink_env::call::Selector::new([0x00; 4])); // balance_of
-        data.push_arg(&accounts.bob);
-        // Push the new execution context to set Bob as caller.
-        ink_env::test::push_execution_context::<ink_env::DefaultEnvironment>(
-            accounts.bob,
-            callee,
-            1000000,
-            1000000,
-            data,
-        );
-
-        // Bob tries to transfer tokens from Alice to Eve.
-        let emitted_events_before =
-            ink_env::test::recorded_events().collect::<Vec<_>>();
-        assert_eq!(
-            erc20.transfer_from(accounts.alice, accounts.eve, alice_balance + 1),
-            Err(Erc20Error::InsufficientBalance)
-        );
-        // Allowance must have stayed the same
-        assert_eq!(
-            erc20.allowance(accounts.alice, accounts.bob),
-            initial_allowance
-        );
-        // No more events must have been emitted
-        let emitted_events_after =
-            ink_env::test::recorded_events().collect::<Vec<_>>();
-        assert_eq!(emitted_events_before.len(), emitted_events_after.len());
-    }
-}
-
-/// For calculating the event topic hash.
-struct PrefixedValue<'a, 'b, T> {
-    pub prefix: &'a [u8],
-    pub value: &'b T,
-}
-
-impl<X> scale::Encode for PrefixedValue<'_, '_, X>
-    where
-        X: scale::Encode,
-{
-    #[inline]
-    fn size_hint(&self) -> usize {
-        self.prefix.size_hint() + self.value.size_hint()
-    }
-
-    #[inline]
-    fn encode_to<T: scale::Output + ?Sized>(&self, dest: &mut T) {
-        self.prefix.encode_to(dest);
-        self.value.encode_to(dest);
+        self._allowances_mut().insert((owner, spender), amount);
+        self.emit_approval_event(owner, spender, amount);
     }
 }
