+++ contracts/token/erc20/lib.rs
@@ -3,3 +3,4 @@
 pub mod impls;
 pub mod traits;
 pub mod stub;
+mod tests;
