+++ contracts/token/erc721/traits.rs
@@ -1,11 +1,9 @@
-use ink_lang as ink;
 use ink_prelude::{string::String, vec::Vec};
-use utils::traits::{AccountId};
+use brush::traits::{AccountId};
 
 pub type Id = [u8; 32];
 
-#[derive(Debug, scale::Encode, scale::Decode, PartialEq, Eq, Clone)]
-#[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
+#[derive(strum_macros::AsRefStr)]
 pub enum Erc721Error {
     Unknown(String),
     CallFailed,
@@ -14,18 +12,11 @@ pub enum Erc721Error {
     TokenExists,
     TokenNotFound,
     CannotInsert,
-    CannotRemove,
     CannotFetchValue,
     NotAllowed,
 }
 
-impl core::fmt::Display for Erc721Error {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        write!(f, "Erc721Error: {:?}", self)
-    }
-}
-
-#[ink::trait_definition]
+#[brush::trait_definition]
 pub trait IErc721 {
     /// Returns the balance of the owner.
     ///
@@ -47,33 +38,22 @@ pub trait IErc721 {
 
     /// Approves or disapproves the operator for all tokens of the caller.
     #[ink(message)]
-    fn set_approval_for_all(&mut self, to: AccountId, approved: bool) -> Result<(), Erc721Error>;
+    fn set_approval_for_all(&mut self, to: AccountId, approved: bool);
 
     /// Approves the account to transfer the specified token on behalf of the caller.
     #[ink(message)]
-    fn approve(&mut self, to: AccountId, id: Id) -> Result<(), Erc721Error>;
+    fn approve(&mut self, to: AccountId, id: Id);
 
     /// Transfer approved or owned token.
     #[ink(message)]
-    fn transfer_from(
-        &mut self,
-        from: AccountId,
-        to: AccountId,
-        id: Id,
-    ) -> Result<(), Erc721Error>;
+    fn transfer_from(&mut self, from: AccountId, to: AccountId, id: Id);
 
     /// Transfers token with `id` from `from` to `to`. Also some `data` can be passed.
     #[ink(message)]
-    fn safe_transfer_from(
-        &mut self,
-        from: AccountId,
-        to: AccountId,
-        id: Id,
-        data: Vec<u8>,
-    ) -> Result<(), Erc721Error>;
+    fn safe_transfer_from(&mut self, from: AccountId, to: AccountId, id: Id, data: Vec<u8>);
 }
 
-#[ink::trait_definition]
+#[brush::trait_definition]
 pub trait IErc721Metadata {
     /// Returns the token name.
     #[ink(message)]
@@ -84,15 +64,15 @@ pub trait IErc721Metadata {
     fn symbol(&self) -> Option<String>;
 }
 
-#[ink::trait_definition]
+#[brush::trait_definition]
 pub trait IErc721Mint {
     /// Creates a new token.
     #[ink(message)]
-    fn mint(&mut self, id: Id) -> Result<(), Erc721Error>;
+    fn mint(&mut self, id: Id);
 
     /// Deletes an existing token. Only the owner can burn the token.
     #[ink(message)]
-    fn burn(&mut self, id: Id) -> Result<(), Erc721Error>;
+    fn burn(&mut self, id: Id);
 }
 
 /// The ERC721Receiver error types.
@@ -100,11 +80,11 @@ pub trait IErc721Mint {
 #[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
 pub enum Erc721ReceiverError {
     /// Returned if transfer is rejected.
-    TransferRejected,
+    TransferRejected(String),
 }
 
 /// Handles the receipt of a single ERC-721 token type.
-#[ink::trait_definition]
+#[brush::trait_definition]
 pub trait IErc721Receiver {
     /// This function is called at the end of a safe_transfer_from after the balance has been updated.
     /// If transfer is rejected it wil return an error.
