--- contracts/utils/lib.rs
@@ -1,4 +1,0 @@
-#![cfg_attr(not(feature = "std"), no_std)]
-
-pub mod macros;
-pub mod traits;
