+++ examples/access-control/Cargo.toml
@@ -15,15 +15,14 @@ ink_prelude = { git = "https://github.com/Supercolony-net/ink", branch = "3.0-rc
 scale = { package = "parity-scale-codec", version = "2.1", default-features = false, features = ["derive"] }
 scale-info = { version = "0.6.0", default-features = false, features = ["derive"], optional = true }
 
-erc721 = { version = "0.1.0-rc2", git = "https://github.com/Supercolony-net/openbrush-contracts", default-features = false, features = ["ink-as-dependency"] }
-access-control = { version = "0.1.0-rc2", git = "https://github.com/Supercolony-net/openbrush-contracts", default-features = false, features = ["ink-as-dependency"] }
-utils = { version = "0.1.0-rc2", git = "https://github.com/Supercolony-net/openbrush-contracts", default-features = false }
+erc721 = { path = "../../contracts/token/erc721", default-features = false, features = ["ink-as-dependency"] }
+access-control = { path = "../../contracts/access/access-control", default-features = false, features = ["ink-as-dependency"] }
+brush = { path = "../../contracts/brush", default-features = false }
 
 [lib]
 name = "my_access_control"
 path = "lib.rs"
 crate-type = [
-	# Used for normal contract Wasm blobs.
 	"cdylib",
 ]
 
@@ -43,6 +42,6 @@ std = [
 
     "erc721/std",
     "access-control/std",
-    "utils/std",
+    "brush/std",
 ]
 ink-as-dependency = []
